Move to a seperate file
